pyproject.toml
[project]
name = "ol-rag-pipeline-core"
version = "0.0.0"
description = "ETL business logic library for OL Enterprise ELT + RAG platform."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "boto3>=1.34.0",
  "httpx>=0.27.0",
  "nats-py>=2.9.0",
  "psycopg[binary]>=3.2.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
